Check for fsc and fsharpc in build.bash and gen-bindings.bash
Default FSharp install on linux installs the compiler and interactive
as fsharpc and fsharpi, not fsc and fsi as on other platforms. This
adds a check to see if FSC is defined and if not checks for fsc and
fsharpc. If no compiler is found/defined it errors and exits the
build.